# Accelerated decompose.nn_filter