#### Reference Issue
Fixes #516


#### What does this implement/fix? Explain your changes.

This PR is a refactor / rewrite of the `nn_filter` operation to avoid using fancy indexing by accessing the sparse matrix index data structure directly.

#### Any other comments?

We get a pretty good speedup over the previous implementation.  Here's a quick benchmark on a 30-second fragment (`nn_filter` is the new version, `librosa.decompose.nn_filter` is the old version):

```python
In [1]: y, sr = librosa.load(librosa.util.example_audio_file(), offset=20, duration=30)
In [2]: S = librosa.amplitude_to_db(librosa.stft(y))
In [3]: R = librosa.segment.recurrence_matrix(S, sparse=True, metric='cosine', mode='affinity')

In [4]: %timeit f1 = nn_filter(S, rec=R, aggregate=np.median)
Out [4]: 1.26 s ± 17.4 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)

In [5]: %timeit f2 = librosa.decompose.nn_filter(S, rec=R, aggregate=np.median)
Out [5]: 5.42 s ± 186 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)

In [6]: %timeit f1 = nn_filter(S, rec=R, aggregate=np.average)
Out [6]: 184 ms ± 11.1 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)

In [7]: %timeit f2 = librosa.decompose.nn_filter(S, rec=R, aggregate=np.average)
Out [7]: 5.28 s ± 227 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
```

<!-- Reviewable:start -->
---
This change is [<img src="https://reviewable.io/review_button.svg" height="34" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/librosa/librosa/598)
<!-- Reviewable:end -->