@bmcfee 2017-07-17
Note: the helper function does not `numba.jit` because it still uses `np.take` (not jittable) as well as helper functions passed by reference (`aggregate`).  The latter could be fixed by doing some scoping tricks to first jit-compile and cache a local function with the hard-coded aggregate op, but the former problem will have to wait until `numba` supports `np.take`.