@bmcfee 2017-07-17
Note: tests pass except for the numba/numpy compat issue on 3.4.