@bmcfee 2017-07-18
Reviewed 1 of 1 files at r1.
Review status: all files reviewed at latest revision, all discussions resolved, some commit checks failed.

---



*Comments from [Reviewable](https://reviewable.io:443/reviews/librosa/librosa/598)*
<!-- Sent from Reviewable.io -->