@bmcfee 2017-07-18
Results are `np.allclose` compatible when using high-precision (`float64` mode), so I'm gonna skip CR on this.